% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significativity_bar.R
\name{significavity_bar}
\alias{significavity_bar}
\title{Significativity bar between 2 groups}
\usage{
significavity_bar(plot, groups, text = "*", text_height = 0.0275,
  size_bar = 1, color_bar = "black", size_text = 8,
  color_text = "black", font_face = 1, font_style = "Arial",
  line_type = "solid")
}
\arguments{
\item{plot}{The ggplot on which you want to draw the bar}

\item{groups}{A vector containing the number (from left to right) of the 2 groups you want to draw the bar between.}

\item{text}{The text (must be a character string) you want on top of the bar. Defaults to "*".}
}
\description{
This function allows you to automatically draw a significativity bar between 2 groups in a ggplot
}
\examples{
significavity_bar(my_plot, c(1, 3), "**")
}
\keyword{ggplot}
